feat(dashboard): Create Box as a centered dialog instead of a right sheet
Few fields no longer warrant a full-height drawer; mirror the API-key
dialog — centered modal, resources laid out in a single CPU/Memory/Disk row.
